# api-docs.json
{"swagger":"2.0","info":{"description":"Api Documentation","version":"1.0","title":"Api Documentation","termsOfService":"urn:tos","contact":{},"license":{"name":"Apache 2.0","url":"http://www.apache.org/licenses/LICENSE-2.0"}},"host":"localhost:8080","basePath":"/","tags":[{"name":"moments-controller","description":"Moments Controller"},{"name":"timelines-controller","description":"Timelines Controller"},{"name":"users-controller","description":"Users Controller"}],"paths":{"/api/moments":{"get":{"tags":["moments-controller"],"summary":"getAll","operationId":"getAllUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/DataResult«List«Moment»»"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/api/moments/add":{"post":{"tags":["moments-controller"],"summary":"add","operationId":"addUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"description","in":"query","description":"description","required":true,"type":"string"},{"name":"files","in":"formData","description":"files","required":false,"type":"array","items":{"type":"file"},"collectionFormat":"multi"},{"name":"moment_date","in":"query","description":"moment_date","required":true,"type":"string","format":"date-time"},{"name":"title","in":"query","description":"title","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Result"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/api/moments/delete":{"post":{"tags":["moments-controller"],"summary":"delete","operationId":"deleteUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"moment","description":"moment","required":true,"schema":{"$ref":"#/definitions/Moment"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Result"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/api/moments/findByMomentDateBetween":{"get":{"tags":["moments-controller"],"summary":"findByMomentDateBetween","operationId":"findByMomentDateBetweenUsingGET","produces":["*/*"],"parameters":[{"name":"From (UTC)","in":"query","description":"From (UTC)","required":true,"type":"string","format":"date-time"},{"name":"To (UTC)","in":"query","description":"To (UTC)","required":true,"type":"string","format":"date-time"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/DataResult«List«Moment»»"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/api/moments/getByTitle":{"get":{"tags":["moments-controller"],"summary":"getByTitle","operationId":"getByTitleUsingGET","produces":["*/*"],"parameters":[{"name":"title","in":"query","description":"title","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/DataResult«Moment»"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/api/moments/update":{"post":{"tags":["moments-controller"],"summary":"update","operationId":"updateUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"moment","description":"moment","required":true,"schema":{"$ref":"#/definitions/Moment"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Result"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/api/moments/{id}":{"get":{"tags":["moments-controller"],"summary":"getById","operationId":"getByIdUsingGET","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/DataResult«Moment»"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/api/timelines":{"get":{"tags":["timelines-controller"],"summary":"getAll","operationId":"getAllUsingGET_1","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/DataResult«List«Timeline»»"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/api/timelines/add":{"post":{"tags":["timelines-controller"],"summary":"add","operationId":"addUsingPOST_1","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"timeline","description":"timeline","required":true,"schema":{"$ref":"#/definitions/Timeline"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Result"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/api/timelines/delete":{"post":{"tags":["timelines-controller"],"summary":"delete","operationId":"deleteUsingPOST_1","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"timeline","description":"timeline","required":true,"schema":{"$ref":"#/definitions/Timeline"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Result"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/api/timelines/getByMomentsIn":{"get":{"tags":["timelines-controller"],"summary":"getByMomentsIn","operationId":"getByMomentsInUsingGET","produces":["*/*"],"parameters":[{"name":"momentId","in":"query","description":"momentId","required":true,"type":"array","items":{"type":"string"},"collectionFormat":"multi"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/DataResult«List«Timeline»»"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/api/timelines/getByTagsIn":{"get":{"tags":["timelines-controller"],"summary":"getByTagsIn","operationId":"getByTagsInUsingGET","produces":["*/*"],"parameters":[{"name":"tags","in":"query","description":"tags","required":true,"type":"array","items":{"type":"string"},"collectionFormat":"multi"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/DataResult«List«Timeline»»"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/api/timelines/getByTitle":{"get":{"tags":["timelines-controller"],"summary":"getByTitle","operationId":"getByTitleUsingGET_1","produces":["*/*"],"parameters":[{"name":"title","in":"query","description":"title","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/DataResult«Timeline»"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/api/timelines/update":{"post":{"tags":["timelines-controller"],"summary":"update","operationId":"updateUsingPOST_1","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"timeline","description":"timeline","required":true,"schema":{"$ref":"#/definitions/Timeline"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Result"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/api/timelines/{id}":{"get":{"tags":["timelines-controller"],"summary":"getById","operationId":"getByIdUsingGET_1","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/DataResult«Timeline»"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/api/users/add":{"post":{"tags":["users-controller"],"summary":"add","operationId":"addUsingPOST_2","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"email","in":"query","required":false,"type":"string"},{"name":"id","in":"query","required":false,"type":"string"},{"name":"password","in":"query","required":false,"type":"string"},{"name":"username","in":"query","required":false,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Result"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}}},"definitions":{"DataResult«List«Moment»»":{"type":"object","properties":{"data":{"type":"array","items":{"$ref":"#/definitions/Moment"}},"message":{"type":"string"},"success":{"type":"boolean"}},"title":"DataResult«List«Moment»»"},"DataResult«List«Timeline»»":{"type":"object","properties":{"data":{"type":"array","items":{"$ref":"#/definitions/Timeline"}},"message":{"type":"string"},"success":{"type":"boolean"}},"title":"DataResult«List«Timeline»»"},"DataResult«Moment»":{"type":"object","properties":{"data":{"$ref":"#/definitions/Moment"},"message":{"type":"string"},"success":{"type":"boolean"}},"title":"DataResult«Moment»"},"DataResult«Timeline»":{"type":"object","properties":{"data":{"$ref":"#/definitions/Timeline"},"message":{"type":"string"},"success":{"type":"boolean"}},"title":"DataResult«Timeline»"},"File":{"type":"object","properties":{"file":{"type":"string","format":"byte"},"fileSize":{"type":"string"},"fileType":{"type":"string"},"filename":{"type":"string"},"id":{"type":"string"}},"title":"File"},"Moment":{"type":"object","properties":{"attachments":{"type":"array","items":{"$ref":"#/definitions/File"}},"creation_date":{"type":"string","format":"date-time"},"description":{"type":"string"},"id":{"type":"string"},"moment_date":{"type":"string","format":"date-time"},"title":{"type":"string"}},"title":"Moment"},"Result":{"type":"object","properties":{"message":{"type":"string"},"success":{"type":"boolean"}},"title":"Result"},"Timeline":{"type":"object","properties":{"creation_date":{"type":"string","format":"date-time"},"description":{"type":"string"},"id":{"type":"string"},"moments":{"type":"array","items":{"$ref":"#/definitions/Moment"}},"tags":{"type":"array","items":{"type":"string"}},"title":{"type":"string"},"user":{"$ref":"#/definitions/User"}},"title":"Timeline"},"User":{"type":"object","properties":{"email":{"type":"string"},"id":{"type":"string"},"password":{"type":"string"},"username":{"type":"string"}},"title":"User"}}}